Resava "The payload is invalid" kada se debuguje pomocu XDebug
By default Laravel will encrypt, and subsequently also decrypt,
all cookies on a request.

When using Xdebug to debug your application from a browser, a
cookie called "XDEBUG_SESSION" is set. As this cookie was not
set, and thus not encrypted, by the Laravel framework, an error
will be thrown when the framework automatically detects and
tries to decrypt the cookie.

The correct solution is to just add the "XDEBUG_SESSION" cookie
to the exceptions array in the App\Http\Middleware\EncryptCookies
middleware.

https://stackoverflow.com/a/49961144